Don't modify actor variables outside of message loop
The deletion code would modify the map of known dirty persistenceIds
outside of the actor loop. This would sometimes blindly overwrite a
Promise in there, so that it would never complete. This, in turn,
resulted in deleteMessages calls never completing after that.

By using a message instead we synchronize access.